backend/requirements.txt
deepface==0.0.79
opencv-python==4.7.0.72
numpy==1.24.3
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
websockets==11.0.3
pillow==9.5.0 